Rearranging the TS code
