+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - The Genesis contract acts as a global register of Chains (chain ID, genesis, oracles, validators)
   - The Namespaces contract acts as a global register where process contracts register themselves
 - The Processes contract has been adapted to the new Namespaces contract
+- `EnsPublicResolver` is now renamed to `EnsResolver`
 
 ## 0.15.0
