--- lib/ens-public-resolver.json
@@ -1,1 +1,0 @@
-../build/ens-public-resolver.json
