+++ lib/ens-resolver.json
@@ -1,0 +1,1 @@
+../build/ens-resolver.json
