+++ lib/genesis.json
@@ -1,0 +1,1 @@
+../build/genesis.json
